% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_generator.R
\name{calculate_length_of_stay}
\alias{calculate_length_of_stay}
\title{Calculate length of stay based on encounter type}
\usage{
calculate_length_of_stay(enc_type)
}
\arguments{
\item{enc_type}{Encounter type code}
}
\value{
Integer length of stay in days
}
\description{
Calculate length of stay based on encounter type
}
