% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcornet.R
\docType{data}
\name{DEFAULT_SOURCES}
\alias{DEFAULT_SOURCES}
\title{Default MPI source systems}
\format{
An object of class \code{list} of length 4.
}
\usage{
DEFAULT_SOURCES
}
\description{
Default MPI source systems
}
\keyword{datasets}
