% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthea_import.R
\name{.load_snomed_icd10_map}
\alias{.load_snomed_icd10_map}
\title{Load SNOMED to ICD-10 mappings for common conditions}
\usage{
.load_snomed_icd10_map()
}
\description{
Load SNOMED to ICD-10 mappings for common conditions
}
\keyword{internal}
