% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthea_import.R
\name{.map_encounter_type}
\alias{.map_encounter_type}
\title{Map Synthea encounter class to PCORnet ENC_TYPE}
\usage{
.map_encounter_type(synthea_class, enc_map)
}
\description{
Map Synthea encounter class to PCORnet ENC_TYPE
}
\keyword{internal}
