% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthea_import.R
\name{.map_snomed_to_icd10}
\alias{.map_snomed_to_icd10}
\title{Map SNOMED code to ICD-10}
\usage{
.map_snomed_to_icd10(snomed_code, snomed_map)
}
\description{
Map SNOMED code to ICD-10
}
\keyword{internal}
