% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_generator.R
\name{generate_encounter_count}
\alias{generate_encounter_count}
\title{Generate number of encounters based on profile}
\usage{
generate_encounter_count(profile_name, profiles = CLINICAL_PROFILES)
}
\arguments{
\item{profile_name}{Name of the clinical profile}

\item{profiles}{List of clinical profile definitions}
}
\value{
Integer number of encounters
}
\description{
Generate number of encounters based on profile
}
