% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_generator.R
\name{generate_profile_labs}
\alias{generate_profile_labs}
\title{Generate lab results for an encounter based on clinical profile}
\usage{
generate_profile_labs(profile, encounter_type)
}
\arguments{
\item{profile}{Profile definition list}

\item{encounter_type}{Type of encounter (IP, ED, AV, etc.)}
}
\value{
Data frame of lab result rows
}
\description{
Generate lab results for an encounter based on clinical profile
}
