% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_generator.R
\name{generate_profile_medications}
\alias{generate_profile_medications}
\title{Generate medications for a patient based on clinical profile}
\usage{
generate_profile_medications(profile)
}
\arguments{
\item{profile}{Profile definition list}
}
\value{
Data frame of medication rows
}
\description{
Generate medications for a patient based on clinical profile
}
