% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_generator.R
\name{generate_profile_procedures}
\alias{generate_profile_procedures}
\title{Generate procedures for an encounter based on clinical profile}
\usage{
generate_profile_procedures(profile, encounter_type)
}
\arguments{
\item{profile}{Profile definition list}

\item{encounter_type}{Type of encounter (IP, ED, AV, etc.)}
}
\value{
Data frame of procedure rows
}
\description{
Generate procedures for an encounter based on clinical profile
}
