% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_generator.R
\name{generate_profile_vitals}
\alias{generate_profile_vitals}
\title{Generate vital signs for an encounter based on clinical profile}
\usage{
generate_profile_vitals(profile)
}
\arguments{
\item{profile}{Profile definition list}
}
\value{
Data frame with single row of vital signs
}
\description{
Generate vital signs for an encounter based on clinical profile
}
