% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcornet.R
\docType{package}
\name{pcornet.synthetic-package}
\alias{pcornet.synthetic}
\alias{pcornet.synthetic-package}
\title{PCORnet CDM Synthetic Database Generator}
\description{
Function-based API for generating synthetic PCORnet CDM and MPI databases
}
\author{
\strong{Maintainer}: PCORnet Synthetic \email{synthetic@example.com}

}
\keyword{internal}
