% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{print_all_tables}
\alias{print_all_tables}
\title{Print First N Rows of All Tables in Both Databases}
\usage{
print_all_tables(con_cdw, con_mpi, n = 10)
}
\arguments{
\item{con_cdw}{Connection to the CDW database}

\item{con_mpi}{Connection to the MPI database}

\item{n}{Number of rows to display per table (default: 10)}
}
\value{
Invisible NULL
}
\description{
Displays a preview of all tables in the CDW and MPI databases.
}
\examples{
\dontrun{
dbs <- create_pcornet_database(n_patients = 50)
print_all_tables(dbs$cdw, dbs$mpi)
}

}
