% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{print_table_summary}
\alias{print_table_summary}
\title{Print Summary of All Tables in Both Databases}
\usage{
print_table_summary(con_cdw, con_mpi)
}
\arguments{
\item{con_cdw}{Connection to the CDW database}

\item{con_mpi}{Connection to the MPI database}
}
\value{
Invisible NULL
}
\description{
Displays row and column counts for all tables in the CDW and MPI databases.
}
\examples{
\dontrun{
dbs <- create_pcornet_database(n_patients = 50)
print_table_summary(dbs$cdw, dbs$mpi)
}

}
