% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_generator.R
\name{select_encounter_type}
\alias{select_encounter_type}
\title{Select encounter type based on clinical profile}
\usage{
select_encounter_type(profile_name)
}
\arguments{
\item{profile_name}{Name of the clinical profile}
}
\value{
Character encounter type code (IP, ED, AV, OA, IS)
}
\description{
Select encounter type based on clinical profile
}
